made it a proper project with more than one compile unit, because aparantly I can't "quick-hack" more than 100 lines of code
