+++ string_helper.hpp
@@ -1,0 +1,9 @@
+
+#ifndef STRINGHELPER_GUARD
+#define STRINGHELPER_GUARD
+
+#include <string>
+
+std::string chomp (const std::string& in);
+
+#endif
